requirements.txt
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.8.2
datasets>=2.14.0
streamlit>=1.30.0
langchain>=0.0.335
langchain-core>=0.1.15
langchain-community>=0.0.1
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
python-docx>=0.8.11
docx2txt>=0.8
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.2 